uploadImages2IG/requirements.txt
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
decorator==4.4.2
future==0.18.2
huepy==1.2.1
idna==2.10
imageio==2.9.0
imageio-ffmpeg==0.4.3
iniconfig==1.1.1
instabot==0.117.0
mock==4.0.3
moviepy==1.0.3
numpy==1.20.1
packaging==20.9
Pillow==8.1.2
pluggy==0.13.1
proglog==0.1.9
py==1.10.0
pyparsing==2.4.7
PySocks==1.7.1
pytest==6.2.2
pytz==2021.1
requests==2.25.1
requests-toolbelt==0.9.1
responses==0.12.1
schedule==1.0.0
scipy==1.6.1
six==1.15.0
toml==0.10.2
tqdm==4.59.0
urllib3==1.26.3
